+++ NEWS.md
@@ -1,6 +1,8 @@
 Version 0.2
+ - enhancement: project support
  - enhancement: paint color editor use system palette's disabled group color
  - fix: add watch not work when there's no editor openned;
  - enhancement: rainbow parenthesis
  - enhancement: run executable with parameters
  - add: widget for function tips
+ - enhancement: options for editor tooltips